requirements-production.txt
fastapi
uvicorn
httpx
python-dotenv
starlette
itsdangerous
PyJWT
python-multipart
cryptography
python-decouple